@@ -4,7 +4,6 @@ description: Demonstrates how to use the url_launcher plugin.
 dependencies:
   flutter:
     sdk: flutter
-  url_launcher: any
   url_launcher_macos:
     # When depending on this package from a real application you should use:
     #   url_launcher_macos: ^x.y.z
@@ -12,17 +11,18 @@ dependencies:
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
     path: ../
+  url_launcher_platform_interface: ^2.0.0
 
 dev_dependencies:
   integration_test:
     path: ../../../integration_test
   flutter_driver:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.8"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.20.0"
